Fix a parser lock for malformed freeform array items
If a freeform array contained an item without a "." indicating that it's a subproperty, Betty would lock in an infinite loop. This isn't technically related to the freeform array: it has to do with the object moving the current scope of the assembler back out of the array (correctly), and then the `closeArray()` function not checking for the output object root when it encounters the (now unnecessary) `[]` tag and tries to backtrack past the nearest array scope.

Weirdly enough, there was no test case in the original NYT set to catch this. I've added a case to the test document and our own automated tests to make sure that it works correctly.